move the logic of promoting or demoting a user to be mod from the userRouter into the modRouter as this action isn't allowed to regular users
